removed navbar from sign in, now need to ensure there is a redirect to the home page after successfully signing in
