+++ hello-world/app/components/NavbarWrapper.tsx
@@ -1,0 +1,11 @@
+"use client";
+
+import { usePathname } from "next/navigation";
+import Navbar from "./Navbar";
+
+export default function NavbarWrapper() {
+  const pathname = usePathname();
+  const showNavbar = pathname !== "/"; // Hide navbar on login page
+
+  return showNavbar ? <Navbar /> : null;
+}
